LayoutTests/fast/device-orientation/device-motion-request-permission-denied.html

<!DOCTYPE html>
<html>
<body>
<script src="../../resources/js-test.js"></script>
<script>
description("Basic testing for DeviceMotionEvent.requestPermission().");
jsTestIsAsync = true;

if (window.testRunner)
    testRunner.setShouldAllowDeviceOrientationAndMotionAccess(false);

addEventListener("devicemotion", () => {});

internals.withUserGesture(() => {
    DeviceMotionEvent.requestPermission().then((_result) => {
        result = _result;
        shouldBeEqualToString("result", "denied");

        addEventListener("devicemotion", () => {});

        internals.withUserGesture(() => {
            DeviceMotionEvent.requestPermission().then((_result) => {
                result = _result;
                shouldBeEqualToString("result", "denied");
                finishJSTest();
            });
        });
    });
});
</script>
</body>
</html>
